bug fix cron

+++ src/modules/mongo/task/crons/history.cron.ts
@@ -1,15 +1,22 @@
 import { Injectable, Logger } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { Cron } from '@nestjs/schedule';
 import { HistoryService } from '../../history/history.service';
 
 @Injectable()
 export class HistoryCron {
   private logger: Logger = new Logger(`Cron - ${HistoryService.name}`);
-  constructor(private _historyService: HistoryService) {}
+  constructor(
+    private _historyService: HistoryService,
+    private configService: ConfigService,
+  ) {}
 
   // delete expired history
   @Cron('0 0 0 * * *')
   async deleteExpiredHistory() {
+    if (this.configService.get('appId') != 'crons') {
+      return;
+    }
     this.logger.log('History expired cron started...');
     const { error } = await this._historyService.$db.deleteBulkRecords({
       options: {
