+++ src/modules/mongo/task/crons/trash.cron.ts
@@ -1,15 +1,22 @@
 import { Injectable, Logger } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { Cron } from '@nestjs/schedule';
 import { TrashService } from '../../trash/trash.service';
 
 @Injectable()
 export class TrashCron {
   private logger: Logger = new Logger(`Cron - ${TrashService.name}`);
-  constructor(private _trashService: TrashService) {}
+  constructor(
+    private _trashService: TrashService,
+    private configService: ConfigService,
+  ) {}
 
   // delete expired trash
   @Cron('0 0 0 * * *')
   async deleteExpiredTrash() {
+    if (this.configService.get('appId') != 'crons') {
+      return;
+    }
     this.logger.log('Trash expired cron started...');
     const { error } = await this._trashService.$db.deleteBulkRecords({
       options: {
